Modify NotificationWindow
